+++ Trabajo Practico 1/FuncionesAritmeticas.h
@@ -3,4 +3,5 @@ int Suma(int numero1,int numero2);
 int Resta(int numero1, int numero2);
 int Multiplicacion(int factor1,int factor2);
 float Division(float dividendo,float divisor);
-long long int Factorial(int primerOperando);
+float Factorial(int primerOperando);
+
